refactor: divide StockManager class into 4 different classes
- FreeAdditionAsker
- NonPromotionAsker
- FreeProductsChecker
- StockDeductor
