+++ src/main/java/store/Application.java
@@ -5,7 +5,10 @@ import store.domain.Products;
 import store.domain.Promotions;
 import store.service.PriceCalculator;
 import store.service.PromotionTimer;
-import store.service.StockManager;
+import store.service.ordermanager.FreeAdditionAsker;
+import store.service.ordermanager.NonPromotionAsker;
+import store.service.stockmanager.FreeProductsChecker;
+import store.service.stockmanager.StockDeductor;
 import store.util.Retrier;
 import store.util.md.MdPaths;
 import store.view.InputView;
@@ -20,10 +23,13 @@ public class Application {
         Retrier retrier = new Retrier();
         PromotionTimer timer = new PromotionTimer(products, promotions);
         PriceCalculator calculator = new PriceCalculator(inputView, products, promotions, timer, retrier);
-        StockManager manager = new StockManager(inputView, products, promotions, timer, retrier);
+        FreeAdditionAsker freeAdditionAsker = new FreeAdditionAsker(inputView, products, promotions, timer, retrier);
+        NonPromotionAsker nonPromotionAsker = new NonPromotionAsker(inputView, products, promotions, timer, retrier);
+        FreeProductsChecker freeProductsChecker = new FreeProductsChecker(products, promotions, timer);
+        StockDeductor deductor = new StockDeductor(products, timer);
 
         StoreController controller = new StoreController(inputView, outputView, products, promotions, retrier, timer,
-                calculator, manager);
+                calculator, freeAdditionAsker, nonPromotionAsker, freeProductsChecker, deductor);
 
         controller.run();
     }
